add confirm delete

+++ public/assets/js/script.js
@@ -43,14 +43,16 @@ $(document).ready(function() {
 
     $(".remove").on("click", function(e) {
         e.preventDefault();
-        var id = $(this).data("id");
+        if (confirm("Confirm delete?")) {
+            var id = $(this).data("id");
 
-        // Send the DELETE request.
-        $.ajax("/api/burgers/" + id, {
-            type: "DELETE"
-        }).then(function() {
-            location.reload();
-        });
+            // Send the DELETE request.
+            $.ajax("/api/burgers/" + id, {
+                type: "DELETE"
+            }).then(function() {
+                location.reload();
+            });
+        }
     });
 
 });
